# faster BigInt hashing