The improvements concerns mainly:
1. ints which fit in one `Limb` (by special casing these more directly)
2. ints which are even: they had to be right shifted in the generic algorithm, leading to allocations.

Here is a small table containing the time for 1 hash (on my laptop) for few different sizes and for even/odd numbers. In parenthesis is the approximate ratio compared to master. "16 (sparse)" means something like `2^1000`, while "16 (dense)" is something like `2^1000+1` or `2^1000+2`.

| num limbs   | master even   | master odd     | PR     |
|-------------|---------------|----------------|--------|
| 1           | 128 ns (18x)  | 39 ns (5.5x)   | 7 ns   |
| 2           | 143 ns (4x)   | 46 ns (1.3x)   | 36 ns  |
| 16 (sparse) | 149 ns (3.5x) | n.a.           | 41 ns  |
| 16 (dense)  | 232 ns (1.8x) | 134 ns (1.08x) | 124 ns |