@StefanKarpinski 2019-11-27
Very clever! I'm happy to merge this whenever.