@rfourquet 2019-12-19
So I reviewed again in detail, still looks good to me :) I had tested with many thousands of values that the old and new versions match.

But I wonder now whether I should revert deleting the `hash_integer(::BitInt, ::UInt)` method: it would not be used anymore in `Base`, but after all this optimized version might be used by custom number types which go through the `hash(x::Real, h::UInt)` method (which in turns calls `hash_integer`), and e.g. have a `decompose` which returns some `BigInt`s.