@JeffreySarnoff 2019-12-19
That is a good thought.  Let's do that and then get this off the runway!