@JeffreySarnoff 2020-02-23
Faster BigInt hashing is ready to be merged.

Working on a custom type, I have run into another (though related) `decompose` error.  The proper fix is that the custom numeric type implement its own type specific method of `decompose`.  It turns out that backstopping `hash_integer(n::BigInt, h::UInt)` is a porous fix (because it was present in the version I used and did not prevent erroring on/through `decompose`).  _I do not have that code_.

The removed function is shown below for reference.
 
```
function hash_integer(n::BigInt, h::UInt)
    s = n.size
    s == 0 && return hash_integer(0, h)
    p = convert(Ptr{UInt}, n.d)
    b = unsafe_load(p)
    h ⊻= hash_uint(ifelse(s < 0, -b, b) ⊻ h)
    for k = 2:abs(s)
        h ⊻= hash_uint(unsafe_load(p, k) ⊻ h)
    end
    return h
end
```