@JeffreySarnoff 2020-02-23
I assumed that not restoring the function indicated a desire to leave it out.  If that is incorrect, so much the better, please put it back [covering paths accessed by reasonable use is reasonable].