@rfourquet 2020-02-24
Sure, I restored `hash_integer(n::BigInt, h::UInt)`. Otherwise, when CI turns green again, this seems good to go for me, no docs or NEWS.md to update, or tests to add.