@JeffreySarnoff 2020-02-24
@StefanKarpinski skies are clear, it is a beautiful day for merging.