@KristofferC 2020-02-24
I didn't understand the reason for keeping `hash_integer(n::BigInt, h::UInt)`. In what cases is it useful and how are a user supposed to know it exists? Is it still tested?