@StefanKarpinski 2020-02-24
IIUC, it's because external code may be relying on it.