@JeffreySarnoff 2020-02-24
I encountered this a few years ago, and the fine grained specifics are a little hazy.   
```
struct ExtendedReal{T} <: Real
     value::T
end
```
Unless `hash` and `decompose` are overloaded to handle this type
using `ExtendedReal` will lean on `hash(x::Real, h::UInt)`.  That calls `hash_integer`, and for `ExtendedReal` the call is to `hash_integer(n::BigInt, h::UInt)`  since `decompose(x::ExtendedReal)` will provide a decomposition that uses `BigInts`.

Writing special purpose `decompose` methods can be difficult to do; the operational semantics of decompose does not always mesh with the nature of an extended real type (e.g. precise yet inexact values, fuzzy numbers).