@rfourquet 2020-02-25
> In what cases is it useful 

To phrase it in a different way than @JeffreySarnoff put it, it's for user-defined integer types which might use the generic `Base` provided `hash(x::Real, h::UInt)`, but overload `decompose` such that one of the components (`num`, `denom`, `pow`) is a `BigInt`. Without this specific `hash_integer(n::BigInt, h::UInt)` method, such type will fall back to using `hash_integer(n::Integer, h::UInt)`, which is quite slower.

> and how is a user supposed to know it exists?

She doesn't have to know. But if we remove it here, while not breaking, she might notice a significant slow-down in hashing her type.

> Is it still tested? 

I believe not. We could add a few tests I guess.

> And if GMP.Limb!== UInt that function definition will just disappear?

Yes, but then the generic slow fall-back will take over.