@KristofferC 2020-02-25
Alright, it just seems unideal to have untested "dead" code in Base "because it can be useful". Anyway, I don't want to hold up this PR but at some point, having a test to make it covered and some comment saying when it is being hit and that it exists as a performance optimization (like in the comment above) would be good imo.