@rfourquet 2020-02-25
> it just seems unideal to have untested "dead" code in Base "because it can be useful".

Yeah I agree, and that's why I deleted it initially. But actually, I was stupid, it's not dead code and is used by `Base`'s own `Rational{BigInt}`:
```julia
julia> @btime hash($(big(2)//3), UInt(0))
  127.409 ns (3 allocations: 48 bytes)               
0x36fc23415080622e

julia> m = collect(methods(Base.hash_integer))[1]    
hash_integer(n::BigInt, h::UInt64) in Base at hashing2.jl:17                                              

julia> Base.delete_method(m)

julia> @btime hash($(big(2)//3), UInt(0))            
  257.855 ns (7 allocations: 128 bytes)              
0x36fc23415080622e                                   
```